pyproject.toml
[project]
name = "animus"
version = "1.6.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "agno>=2.5.8",
  "alembic>=1.18.4",
  "faker>=40.4.0",
  "fastapi[standard]>=0.128.7",
  "google-cloud-storage>=3.4.1",
  "inngest>=0.5.15",
  "pypdf>=6.1.3",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "pwdlib[argon2]>=0.3.0",
  "pyjwt>=2.11.0",
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "python-docx>=1.2.0",
  "python-dotenv>=1.0.0",
  "sqlalchemy>=2.0.46",
  "python-ulid",
  "testcontainers>=4.14.1",
  "psycopg2-binary>=2.9.11",
  "python-jose>=3.5.0",
  "redis>=5.0.0",
  "resend>=2.13.0",
  "scalar-fastapi>=1.8.1",
  "google-auth>=2.49.1",
  "google-cloud-aiplatform>=1.143.0",
  "google-genai>=1.68.0",
  "qdrant-client>=1.17.1",
  "google-genai>=1.68.0",
  "sentence-transformers>=5.3.0",
  "openai>=2.30.0",
]

[dependency-groups]
dev = [
  "basedpyright>=1.38.2",
  "poethepoet>=0.41.0",
  "pyright>=1.1.408",
  "pytest>=9.0.2",
  "pytest-mock>=3.15.1",
  "ruff>=0.15.0",
]

[project.scripts]
dev = "main:main"

[tool.poe.tasks."dev"]
cmd = "uv run dev"

[tool.poe.tasks."pubsub"]
cmd = "npx --ignore-scripts=false inngest-cli@latest dev -u http://localhost:8080/api/inngest"

[tool.poe.tasks."db:migrate"]
cmd = "alembic revision --autogenerate -m \"${message}\""
args = [{ name = "message", positional = true }]

[tool.poe.tasks."db:upgrade"]
cmd = "alembic upgrade ${target}"
args = [{ name = "target", positional = true, default = "head" }]

[tool.poe.tasks."db:downgrade"]
cmd = "alembic downgrade ${target}"
args = [{ name = "target", positional = true, default = "-1" }]

[tool.poe.tasks."db:sync"]
cmd = "alembic stamp head"

[tool.poe.tasks."db:current"]
cmd = "alembic current"

[tool.poe.tasks."db:seed"]
cmd = "uv run python -c \"from animus.database.sqlalchemy.seed import seed; seed()\""

[tool.poe.tasks.test]
cmd = "pytest -s -x -vv"

[tool.poe.tasks.typecheck]
cmd = "basedpyright"

[tool.poe.tasks.codecheck]
shell = "ruff check --fix src tests && ruff format --check src tests"


[tool.ruff]
line-length = 88
target-version = "1.6.0"
fix = true
show-fixes = true
indent-width = 4
exclude = [
  "venv",
  ".venv",
  "env",
  ".env",
  "node_modules",
  "__pycache__",
  "alembic",
]

[tool.ruff.lint]
select = [
  "ASYNC",
  "A",
  "ANN",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FBT",
  "FIX",
  "I",
  "ICN",
  "ISC",
  "RET",
  "RSE",
  "S",
  "SIM",
  "SLF",
  "T10",
  "T20",
  "TC",
  "TD",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
  "RUF",
  "N",
]
ignore = [
  "T201",
  "COM812",
  "ANN201",
  "I001",
  "Q000",
  "RUF022",
  "TRY003",
  "EM102",
  "E501",
  "A002",
  "EM101",
  "F821",
  "FBT001",
  "FBT002",
  "ANN205",
  "ANN401",
  "B008",
  "RUF006",
  "S105",
]


[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["animus"]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
include = ["src", "tests"]
stubPath = "stubs"
reportMissingTypeStubs = false
exclude = [
  "**/venv",
  "**/.venv",
  "**/env",
  "**/.env",
  "**/node_modules",
  "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = [".", "src"]
testpaths = ["tests"]
filterwarnings = [
  "ignore::pydantic.warnings.PydanticDeprecatedSince20:inngest.*",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
